Refactor GroupStats with container
